Add trailing slash for status-updates.
The URL for updating the status-update resource was being formatted
without a trailing slash. Review Board 3.x would properly redirect this,
but a change in the version of Django that 4.x uses broke this and would
result in a 404. In either case, it's better to always include the
trailing slash to avoid excess requests.

Reviewed at https://reviews.reviewboard.org/r/11718/