Introduce downloadOnly and downloadRbt
This change introduces two new configuration keys for tweaking the way
jobs are run.

The "installRBTools" configuration key will include a step for
installing the RBTools package. This can be unchecked in the case where
the admin has installed a system-wide RBTools package already.

The "downloadOnly" configuration key will cause the plugin to use
RBTools to download the patch, but skip applying it. This allows the
user more control about how to apply the patch in special cases.

Testing Done:
Added plugin to local jenkins and tested use of the new keys.

Also added unit tests for it.

Reviewed at https://reviews.reviewboard.org/r/11677/